+++ dsapy/gui/sorting_visualizer/sorting_config.py
@@ -1,0 +1,22 @@
+# SIZE
+SIZE_CONFIG = {
+    'init_val': 100,
+    'from': 50,
+    'to': 1000
+}
+
+# --- ALGORITHM ---
+ALGORITHM_CONFIG = {
+    'algorithms': [
+        'Bubble Sort',
+        'Selection Sort',
+        'Insertion Sort'
+    ]
+}
+
+# --- DELAY ---
+DELAY_CONFIG = {
+    'init_val': 5,
+    'from': 0,
+    'to': 10
+}