+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## [0.0.2] - 2022-11-26
 
 ### Added
+- insertion_sort.py module and tests
+- selection_sort.py module and tests
 - bubble_sort.py module and tests
 - sorting algorithms package