+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [0.0.3] - 2022-12-01
+
+### Added
+- Sorting Visualizer GUI
+
 ## [0.0.2] - 2022-11-26
 
 ### Changed